Move AVM circuit builder
