+++ barretenberg/cpp/src/barretenberg/vm/avm_trace/avm_common.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "barretenberg/proof_system/circuit_builder/circuit_builder_base.hpp"
-#include "barretenberg/proof_system/circuit_builder/generated/avm_circuit_builder.hpp"
+#include "barretenberg/vm/generated/avm_circuit_builder.hpp"
 #include <cstdint>
 
 namespace bb::avm_trace {
